add a basic tui which loads the front page
